requirements.txt
langchain
langchain-community
langchain-huggingface
python-dotenv
requests